feat: tcpinfo on peers
